+++ src/lib.rs
@@ -37,6 +37,7 @@
 //! usable for my personal projects, and the number of working tests are
 //! improving.
 #![forbid(unsafe_code)]
+// TODO: #![forbid(missing_docs)]
 use std::path::Path;
 
 pub mod css;
@@ -48,7 +49,7 @@ pub mod output;
 mod parser;
 pub mod sass;
 pub mod selectors;
-mod value;
+pub mod value;
 mod variablescope;
 
 pub use crate::error::Error;
@@ -59,12 +60,7 @@ pub use crate::parser::{
     parse_scss_data, parse_scss_file, parse_scss_path, parse_value_data,
     ParseError, SourcePos,
 };
-pub use crate::sass::Item;
-pub use crate::value::{
-    colors, Dimension, ListSeparator, Number, Quotes, Unit,
-};
 pub use crate::variablescope::{GlobalScope, Scope};
-pub use num_rational::Rational;
 
 /// Parse a scss value from a buffer and write its css representation
 /// in the given format.