+++ CHANGELOG.md
@@ -14,6 +14,7 @@ project adheres to
 * Improve name lookups in scopes and modules, PR #87.
 * A color can be Hsla or Hwba as well as Rgba. PR #88, #89.
 * Handle units in `@for` loops inside sass functions.
+* Some types moved into `value` mod, imporoved docs. PR #90.
 * Test suite sass-spec updated to 2021-02-04.
 * Some cleanups.